pseudoattr/requirements_ext.txt
torch==1.12.1
torchaudio==0.12.1
pandas==1.4.2
umap-learn==0.5.6
tqdm==4.66.5
numpy==1.23.5
torchopenl3
panns-inference
